Move the transport code to a dedicated module.
